13_IntroductionToDjangoRESTFramework/djlibrary/app_libs/serializers.py
from rest_framework import serializers
from app_libs.models import Author, Book


class AuthorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Author
        fields = ['id', 'name', 'family', 'birth_year']


class BookSerializer(serializers.ModelSerializer):
    class Meta:
        model = Book
        fields = ['id', 'title', 'isbn', 'year', 'pages']
